requirements.txt
# Core dependencies
asyncio
dataclasses
typing-extensions
pathlib

# Data processing
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=10.0.0
openpyxl>=3.0.0
chardet>=5.0.0

# Database connectivity
sqlalchemy>=1.4.0
pymysql>=1.0.0
psycopg2-binary>=2.9.0
snowflake-connector-python>=2.8.0

# Cloud storage
azure-storage-blob>=12.0.0
boto3>=1.26.0
azure-identity>=1.12.0

# HTTP/API
aiohttp>=3.8.0
requests>=2.28.0
httpx>=0.24.0

# Configuration and environment
python-dotenv>=0.19.0
pydantic>=1.10.0

# Logging and monitoring
structlog>=22.0.0
rich>=12.0.0

# Databricks integration
databricks-sql-connector>=2.0.0
databricks-cli>=0.17.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional dependencies for enhanced functionality
# Uncomment as needed for specific use cases

# Jupyter notebook support
# jupyter>=1.0.0
# ipywidgets>=8.0.0

# Machine learning (for data quality assessment)
# scikit-learn>=1.1.0
# scipy>=1.9.0

# Advanced data validation
# cerberus>=1.3.0
# jsonschema>=4.0.0

# Performance monitoring
# psutil>=5.9.0
# memory-profiler>=0.60.0
